Add gtp support
NOTE: We implement victiplay with gtp in a slightly hacky way. We
require that the user specify which color the gtp bot will be run as
when the `katago` executable is called. The color should not be changed
after this, even though it is technically allowed by the gtp
specification. If you want to play as both colors, you need to launch
two different versions of the gtp bot.