Implement victimplay (#1)
* Start on victimplay.cpp

Currently victimplay.cpp is just a copy of selfplay.cpp. The
purpose of creating this file is so that I can better track
what modfications are made during editing.

* Basic victimplay working?

TODO: Need to modify training data writing to not write victim
data. We don't want to train on victim data because if the victim
is really weak, this will cause the adversary to try to learn the
victim's weak policy, which is not what we want.

* Don't write data for victim in victimplay

* Log victimplay game scores

* Even better logging

* Fix victimplay nnEval swapping

Change so that only adversary gets swapped out, NOT victim.

* Code review

* Small refactor for clarity

* First pass at implementing nneval_colored

* nneval_colored compiles

* Fix bugs

* Fold victimplay logic into selfplay.cpp

* Update victimplay1.cfg

* Add comment for assert

* Refactor selfplay and fix bug

* Adjust victimplay1.cfg and add comments

* Add comment about clearBotBeforeSearchThisGame

* Update cpp/configs/training/victimplay1.cfg

Co-authored-by: Adam Gleave <adam@gleave.me>

* Update cpp/dataio/trainingwrite.cpp

Co-authored-by: Adam Gleave <adam@gleave.me>

* Fix NNEvaluator::serve

* Add victimplay capabilities for match and gtp (#3)

* Add victimplay capabilities for match

* Adjust supportsShorttermError()

* Add gtp support

NOTE: We implement victiplay with gtp in a slightly hacky way. We
require that the user specify which color the gtp bot will be run as
when the `katago` executable is called. The color should not be changed
after this, even though it is technically allowed by the gtp
specification. If you want to play as both colors, you need to launch
two different versions of the gtp bot.

* Fix a bug and warning

* Refactor nneval stuff

* Refactor control flow

* Better comment regarding checkForNewNNEval

* Fix trainingwrite for victimplay

Co-authored-by: Adam Gleave <adam@gleave.me>