Save work for MCTS playout tests
TL;DR Not working, but we are not stuck!

Working on tests for MCTS playouts at the moment. My reference
implementation is not matching at the moment. But I have a plan to debug
it. We will just print out the tree and see where the search goes awry.

A even more involved backup solution is to modify the katago search code
to print out intermediate search progress.