Improve EMCTS1 implementation
Now if we visit an opponent node for the first time, we will immediately
expand that node again. This avoids having dangling nodes in the tree
with 0 weight, which are a waste of compute and also might screw with
the implementation in weird ways.

We also implement tests for this improved EMCTS1 implementation!
We have tests for both the backup and the forward playout process!