Customizable passing behaviors (#25)
* Improve loadParams for EMCTS1

* Add constant policy networks (and tests for them)

* Fix typing

* Major refactor of victimplay

Remove NNEval_Colored and replace with custom logic inside search.

* Fix a weightSum <= 0.0 condition

Not sure why this was not necessary before the refactor...

* Wrote test for MCTS backup

* Save work for MCTS playout tests

TL;DR Not working, but we are not stuck!

Working on tests for MCTS playouts at the moment. My reference
implementation is not matching at the moment. But I have a plan to debug
it. We will just print out the tree and see where the search goes awry.

A even more involved backup solution is to modify the katago search code
to print out intermediate search progress.

* Both MCTS and EMCTS1 backups are OK!

Still need to check forward playout process.

* Fix oppNNEval bug

* MCTS playout check working!

* Refactor test code a bit

TODO: Write EMCTS1 playout test

* Implemented various passing behaviors

* Hacky fix for error when opponent has no legal moves

* Improve EMCTS1 implementation

Now if we visit an opponent node for the first time, we will immediately
expand that node again. This avoids having dangling nodes in the tree
with 0 weight, which are a waste of compute and also might screw with
the implementation in weird ways.

We also implement tests for this improved EMCTS1 implementation!
We have tests for both the backup and the forward playout process!

* Change to C++17

So we can use structure bindings:
https://stackoverflow.com/a/40914071/1337463

* Add tests for final move selection

* Make number of moves and visits configurable

Set default github ci to run a smaller test.

* Fix error in AvoidPassAliveTerritory passing behavior

* Fixed avoid-pass-alive-territory when maxVisits=1

* Sorted some config statements

* Remove -u flag from /bin/bash in shell scripts

* Remove -eu flag in shell scripts for K8 compat

* Reverting more stupid autoformatting

* Even more reverting of stupid autoformatting

* Eradicated last of autoformatting

* Fixed uninitialized field bug in SearchParams

* Replace all instances of unsafe uninitialized use of SearchParams

* Add extra newline

Co-authored-by: Adam Gleave <adam@gleave.me>

* Reverting superfluous constructor calls in C++

* Fix NoSuicide logic

* Make darglint happy

* Incorporate Tony's suggestions for code clarity

Co-authored-by: Tony Wang <twang6@mit.edu>
Co-authored-by: ttw <1733389+ed1d1a8d@users.noreply.github.com>
Co-authored-by: Adam Gleave <adam@gleave.me>