EMCTS1.1: Perfect victim simulation (#34)
### Summary
Improves EMCTS1 so that the adversary can model the victim perfectly (modulo subtle randomness in the MCTS search procedure itself). Gets rid of old code that did approximate visit=1 modeling (approximate because many victim configuration settings were not factored in).

We accomplish perfect victim modeling by having the adversary have a `Search` object that corresponds to the victim. The adversary has the option of simulating the victim with a lower visit count via a few override flags (which recovers the old behavior of EMCTS1).

We also reverted the `nnEvaluator` member variable back to how it was originally used in the codebase.