victimplay: Print victim config reload only once, not once for each thread (#79)
PR #74 updated the victimplay victim config reloading logic so that it could be reused for the gatekeeper. But I had a bug where if the config got updated, every game thread would try to load the config and print a message about the config getting updated. We only need one thread to load the config, and we only want to print the message once. (I thought I fixed this in PR #74, but maybe I forgot to push a commit.)

Changes:
* Only the first thread to notice the config change will load the config into `paramss` and print a message about the config getting updated