train.py: Add -disable-vtimeloss to PyTorch training (#98)
With the old TF training code we had a flag `-disable-vtimeloss` to disable the variance time loss. We removed it when we moved to KataGo v1.12.0, which changes the training code to PyTorch, since we didn't want to add extraneous changes to the PyTorch code and we weren't using the PyTorch code.

This PR adds the `-disable-vtimeloss` flag back.

Testing: Checked that running `train.py` with `-disable-vtimeloss` drives the loss down without driving vtimeloss down.